+++ backend/src/jv2backend/io/tests/test_isisJournalServer.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (c) 2023 Team JournalViewer and contributors
 
-from jv2backend.journalClasses  import Journal
+from jv2backend.journals  import Journal
 from jv2backend.io.isis.isisJournalServer import ISISJournalServer
 
 import pytest
